+++ eslint.config.js
@@ -25,9 +25,6 @@ export default tseslint.config(
             ],
             '@typescript-eslint/no-unused-vars': [
                 'warn',
-                {
-                    argsIgnorePattern: '^_',
-                },
             ],
         },
     },
