+++ src/features/auth/hooks/useSocialLogin.ts
@@ -6,7 +6,7 @@ import {
     clearGooglePopup,
 } from 'src/features/auth/utils/popupManager'
 import { RootState } from 'src/global/store/redux/reduxStore'
-import { authAction } from '../store/authReducer'
+import { authAction } from '../authReducer.ts'
 import {
     useAppDispatch,
     useAppSelector,
